全文検索関数を切り出し
